requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
openai==1.51.0
python-dotenv==1.0.0
aiofiles==24.1.0
httpx==0.27.0